AI-Part/requirements.txt
livekit-agents
python-dotenv
openai
assemblyai
livekit-plugins-assemblyai
livekit-plugins-openai
livekit-plugins-noise-cancellation
livekit-plugins-silero
livekit-plugins-google
langdetect
fastapi
uvicorn[standard]
livekit